Move IDbContext to its own class
