+++ Mehdime.Entity/Interfaces/IAmbientDbContextLocator.cs
@@ -7,10 +7,6 @@
  */
 
 using System;
-using System.Data.Entity;
-using System.Data.Entity.Infrastructure;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Mehdime.Entity
 {
@@ -25,15 +21,5 @@ namespace Mehdime.Entity
         ///     Otherwise returns null.
         /// </summary>
         TDbContext Get<TDbContext>() where TDbContext : class, IDbContext;
-    }
-
-    public interface IDbContext : IDisposable
-    {
-        DbContextConfiguration Configuration { get; }
-        Database Database { get; }
-
-        int SaveChanges();
-
-        Task<int> SaveChangesAsync(CancellationToken cancelToken);
     }
 }
