+++ Mehdime.Entity/Interfaces/IDbContext.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Data.Entity;
+using System.Data.Entity.Infrastructure;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Mehdime.Entity
+{
+    public interface IDbContext : IDisposable
+    {
+        DbContextConfiguration Configuration { get; }
+        Database Database { get; }
+
+        int SaveChanges();
+
+        Task<int> SaveChangesAsync(CancellationToken cancelToken);
+    }
+}
