+++ Mehdime.Entity/Mehdime.Entity.csproj
@@ -56,6 +56,7 @@
     <Compile Include="Implementations\DbContextScopeFactory.cs" />
     <Compile Include="Enums\DbContextScopeOption.cs" />
     <Compile Include="Interfaces\IAmbientDbContextLocator.cs" />
+    <Compile Include="Interfaces\IDbContext.cs" />
     <Compile Include="Interfaces\IDbContextCollection.cs" />
     <Compile Include="Interfaces\IDbContextFactory.cs" />
     <Compile Include="Interfaces\IDbContextReadOnlyScope.cs" />
